上车调试改了一堆东西 懒得一个一个列了

+++ arduino_line_follower/src/DelayChecker.hpp
@@ -4,10 +4,11 @@
 #include <Arduino.h>
 
 #include "MyBitset.hpp"
+// #include "SerialIO.hpp"
 
 template <unsigned N> class DelayChecker {
  private:
-    int duration[N], targetTime[N];
+    int32_t duration[N], targetTime[N];
     char trigger[N];
     MyBitset<N> flags;
 
@@ -15,11 +16,11 @@ template <unsigned N> class DelayChecker {
     void init() {
         for (int i = 0; i < N; ++i) trigger[i] = -1;
     }
-    void setDuration(int i, int time_ms) { duration[i] = time_ms; }
-    void bind(int i, int j) { trigger[i] = j; }
-    template <typename... T> void bind(int i, int j, T... t) { bind(i, j), bind(j, t...); }
-    void bindDiscard(int i, int j) { trigger[i] = j + N; }
-    void discard(int i) {
+    void setDuration(int32_t i, int32_t time_ms) { duration[i] = time_ms; }
+    void bind(int32_t i, int32_t j) { trigger[i] = j; }
+    template <typename... T> void bind(int32_t i, int32_t j, T... t) { bind(i, j), bind(j, t...); }
+    void bindDiscard(int32_t i, int32_t j) { trigger[i] = j + N; }
+    void discard(int32_t i) {
         if (flags.check(i)) {
             flags.reset(i);
             if (trigger[i] != -1) {
@@ -31,16 +32,23 @@ template <unsigned N> class DelayChecker {
     }
     void update() {
         for (int i = 0; i < N; ++i)
-            if (flags[i] && millis() - targetTime[i] >= 0) discard(i);
+            // if (flags[i] && (int32_t)millis() - targetTime[i] >= 0) {
+            //     SerialIO::write(millis(), targetTime[i]);
+            //     discard(i);
+            // } else {
+            //     SerialIO::write(0, 0);
+            // }
+            if (flags[i] && (int32_t)millis() - targetTime[i] >= 0) discard(i);
+        // SerialIO::flush();
     }
-    bool check(int i) const { return flags[i]; }
-    bool operator[](int i) const { return check(i); }
+    bool check(int32_t i) const { return flags[i]; }
+    bool operator[](int32_t i) const { return check(i); }
     char get() {
         for (int i = 0; i < N; ++i)
             if (flags[i]) return i;
         return -1;
     }
-    void set(int i) {
+    void set(int32_t i) {
         flags.set(i);
         targetTime[i] = millis() + duration[i];
     }
